improve hygiene of vpg implementation

+++ spinup/algos/mytorch/vpg/core.py
@@ -73,7 +73,7 @@ class MLPCategoricalActor(Actor):
         self._mlp = mlp([obs_dim] + list(hidden_sizes) + [act_dim], activation)
 
     def _distribution(self, obs):
-        logits = self._mlp.forward(obs)
+        logits = self._mlp.forward(obs.float())
         return Categorical(logits=logits)
 
     def _log_prob_from_distribution(self, pi, act):
@@ -99,7 +99,7 @@ class MLPCritic(nn.Module):
         self._mlp = mlp([obs_dim] + list(hidden_sizes) + [1], activation)
 
     def forward(self, obs):
-        return self._mlp.forward(obs)
+        return self._mlp.forward(obs.float())
 
 class MLPActorCritic(nn.Module):
